requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fsspec==2025.5.1
future==1.0.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==25.0
regex==2024.11.6
requests==2.32.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
yt-dlp==2025.5.22
zipp==3.22.0
